pyproject.toml
[project]
name = "langtrace-tracking"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Denis Ergashbaev", email = "denis.ergashbaev@gmail.com"},
]
dependencies = [
    "dspy-ai>=2.4.12",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "pydantic>=2.8.2",
    "langtrace-python-sdk>=2.2.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
